refactor: 전체 컴포넌트 구조 개선 및 컴포넌트별 스타일 분리
