+++ src/components/common/DeleteItemButton/DeleteItemButton.style.ts
@@ -1,0 +1,14 @@
+import styled from '@emotion/styled';
+
+export const DeleteItemButton = styled.button`
+  height: 24px;
+  border: 1px solid #e5e5e5;
+  border-radius: 4px;
+  padding: 0 8px;
+
+  color: #0a0d13;
+  font-size: 12px;
+  font-weight: 500;
+
+  cursor: pointer;
+`;
