+++ src/components/common/DeleteItemButton/DeleteItemButton.tsx
@@ -1,0 +1,9 @@
+import * as Styled from './DeleteItemButton.style';
+
+interface DeleteItemButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  buttonText: string;
+}
+
+export default function DeleteItemButton({ buttonText, ...props }: DeleteItemButtonProps) {
+  return <Styled.DeleteItemButton {...props}>{buttonText}</Styled.DeleteItemButton>;
+}
