+++ src/components/common/NavigationBar/NavigationBar.style.ts
@@ -1,0 +1,16 @@
+import styled from '@emotion/styled';
+
+export const NavigationBar = styled.header`
+  width: 100%;
+  height: 64px;
+  padding: 0 24px;
+  background: #000000;
+
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+
+  color: #ffffff;
+  font-size: 20px;
+  font-weight: 800;
+`;
