+++ src/components/common/NavigationBar/NavigationBar.tsx
@@ -1,0 +1,6 @@
+import { PropsWithChildren } from 'react';
+import * as Styled from './NavigationBar.style';
+
+export default function NavigationBar({ children, ...props }: PropsWithChildren) {
+  return <Styled.NavigationBar {...props}>{children}</Styled.NavigationBar>;
+}
