+++ src/components/common/PageTitle/PageTitle.style.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const PageTitle = styled.h2`
+  font-size: 24px;
+  font-weight: 700;
+  color: #000000;
+`;
