+++ src/components/common/PageTitle/PageTitle.tsx
@@ -1,0 +1,9 @@
+import * as Styled from './PageTitle.style';
+
+interface PageTitleProps {
+  title: string;
+}
+
+export default function PageTitle({ title }: PageTitleProps) {
+  return <Styled.PageTitle>{title}</Styled.PageTitle>;
+}
