+++ src/components/common/index.ts
@@ -1,0 +1,6 @@
+export { default as CheckBox } from './CheckBox/CheckBox';
+export { default as DeleteItemButton } from './DeleteItemButton/DeleteItemButton';
+export { default as FooterButton } from './FooterButton/FooterButton';
+export { default as NavigationBar } from './NavigationBar/NavigationBar';
+export { default as PageTitle } from './PageTitle/PageTitle';
+export { default as PreviousPageButton } from './PreviousPageButton/PreviousPageButton';
