+++ src/components/shoppingCart/CartContainer/CartContainer.style.ts
@@ -1,0 +1,15 @@
+import styled from '@emotion/styled';
+
+export const EmptyCartContainer = styled.div({
+  width: '100%',
+  height: '100%',
+  flex: '1 0 auto',
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  justifyContent: 'center',
+
+  color: '#0A0D13',
+  fontSize: '16px',
+  fontWeight: '400',
+});
