+++ src/components/shoppingCart/CartItemContainer/CartItemContainer.style.ts
@@ -1,0 +1,24 @@
+import styled from '@emotion/styled';
+
+export const Description = styled.h3({
+  fontSize: '12px',
+  fontWeight: '500',
+  color: '#0A0D13',
+  marginTop: '12px',
+});
+
+export const CheckAllBoxContainer = styled.div({
+  height: '24px',
+  display: 'flex',
+  flexDirection: 'row',
+  gap: '8px',
+  alignItems: 'center',
+  color: '#0A0D13',
+  fontSize: '12px',
+  fontWeight: '500',
+  marginBottom: '20px',
+});
+
+export const CartItemListContainer = styled.ul({
+  margin: '36px 0',
+});
