+++ src/components/shoppingCart/index.ts
@@ -1,0 +1,5 @@
+export { default as CartContainer } from './CartContainer/CartContainer';
+export { default as CartItem } from './CartItem/CartItem';
+export { default as CartItemContainer } from './CartItemContainer/CartItemContainer';
+export { default as OrderAmount } from './OrderAmount/OrderAmount';
+export { default as QuantityController } from './QuantityController/QuantityController';
