+++ src/constants/endpoints.ts
@@ -1,6 +1,4 @@
-const ENDPOINTS = {
+export const ENDPOINTS = {
   shoppingCart: '/react-shopping-cart',
   confirmOrder: '/react-shopping-cart/confirm-order',
 } as const;
-
-export default ENDPOINTS;
