+++ src/page/ShoppingCartPage/ShoppingCartPage.style.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+export const CartContent = styled.section`
+  padding: 36px 24px;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  flex: 1 0 auto;
+`;
