+++ src/page/index.ts
@@ -1,0 +1,2 @@
+export { default as ConfirmOrderPage } from './ConfirmOrderPage/ConfirmOrderPage';
+export { default as ShoppingCartPage } from './ShoppingCartPage/ShoppingCartPage';
