+++ src/utils/convertToLocaleAmount.ts
@@ -1,0 +1,3 @@
+export function convertToLocaleAmount(amount: number) {
+  return amount.toLocaleString('ko-KR').concat('원');
+}
