--- src/utils/convertToLocalePrice.ts
@@ -1,3 +1,0 @@
-export default function convertToLocaleAmount(amount: number) {
-  return amount.toLocaleString('ko-KR').concat('원');
-}
